requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.13
django-ckeditor==6.4.2
django-environ==0.8.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
